Refactor testing infrastructure.
