+++ numba/cuda/tests/cudapy/test_transpose.py
@@ -2,7 +2,7 @@ import numpy as np
 from numba import cuda
 from numba.cuda.kernels.transpose import transpose
 from numba.cuda.testing import unittest
-from numba.tests.ddt import ddt, data, unpack
+from numba.testing.ddt import ddt, data, unpack
 from numba.cuda.testing import skip_on_cudasim
 
 @skip_on_cudasim('Device Array API unsupported in the simulator')
