+++ numba/testing/__main__.py
@@ -1,0 +1,4 @@
+import sys
+from . import run_tests
+
+sys.exit(0 if run_tests(sys.argv).wasSuccessful() else 1)
