+++ numba/tests/__init__.py
@@ -6,6 +6,7 @@ import time
 import warnings
 from unittest.suite import TestSuite
 from numba.testing import load_testsuite
+from numba.testing import ddt # for backward compatibility
 
 
 try:
