+++ numba/tests/npyufunc/__init__.py
@@ -1,0 +1,9 @@
+from unittest.suite import TestSuite
+from numba.testing import load_testsuite
+from os.path import dirname
+
+def load_tests(loader, tests, pattern):
+    suite = TestSuite()
+    suite.addTests(load_testsuite(loader, dirname(__file__)))
+    return suite
+
